+++ src/components/features/profile/ProfileClient/ProfileClient.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 
-import PostList from '@/components/features/posts/PostList/PostList';
+import PostList from '@/features/posts/components/PostList/PostList';
 import { Avatar } from '@/components/ui/Avatar/Avatar';
 import { TimelineContext, type TimelineContextType } from '@/contexts/TimelineContext';
 import { PostWithProfile } from '@/types';
