+++ src/features/posts/components/LikeButton/LikeButton.module.scss
