+++ src/features/posts/components/PostCard/PostCard.module.scss
